Cleanup data output datset association schemas
