+++ backend/app/data_outputs_datasets/router.py
@@ -5,7 +5,7 @@ from sqlalchemy.orm import Session
 
 from app.core.auth.auth import get_authenticated_user
 from app.core.authz import Action, Authorization, DataOutputDatasetAssociationResolver
-from app.data_outputs_datasets.schema import DataOutputDatasetAssociation
+from app.data_outputs_datasets.schema_get import DataOutputDatasetAssociationsGet
 from app.data_outputs_datasets.service import DataOutputDatasetService
 from app.database.database import get_db_session
 from app.users.schema import User
@@ -80,5 +80,5 @@ def remove_data_output_link(
 def get_user_pending_actions(
     db: Session = Depends(get_db_session),
     authenticated_user: User = Depends(get_authenticated_user),
-) -> list[DataOutputDatasetAssociation]:
+) -> list[DataOutputDatasetAssociationsGet]:
     return DataOutputDatasetService().get_user_pending_actions(db, authenticated_user)
