+++ backend/app/data_outputs_datasets/schema_basic.py
@@ -4,7 +4,7 @@ from app.data_outputs_datasets.enums import DataOutputDatasetLinkStatus
 from app.shared.schema import ORMModel
 
 
-class BaseDataOutputDatasetAssociationGet(ORMModel):
+class DataOutputDatasetAssociationBasic(ORMModel):
     id: UUID
     dataset_id: UUID
     data_output_id: UUID
